refactor synced player list to be slightly more future-proof
change this to not iterate over the entire player list every frame, but
instead build a cache of the player list every 5 ticks

this may result in an incredibly minor performance improvement, but is
primarily intended to make this ready for a future game/render thread
split.

also added contributor name colors while i was here, because it seemed
like a sensible idea.